rewrite agent e2e test tasks
